+++ R/get.R
@@ -8,15 +8,22 @@
 #' @param table Name of a table within the mart, otherwise `NULL` (the default)
 #' @param xmart_server Either 'UAT' (the default) or 'PROD'
 #' @param top Number of rows of a table to return
-#' @param query A single string fitting the [Odata protocol](https://www.odata.org/documentation/odata-version-2-0/uri-conventions/) that must start with \code{"$filter="}.
+#' @param query A single string fitting the [Odata protocol](https://www.odata.org/documentation/odata-version-2-0/uri-conventions/)
+#'     that must start with \code{"$filter="}.
+#' @param format Specification of the output format to be returned by the xMart API.
+#'     Defaults to `"none"`, but consider switching to `"csv"` if you are having
+#'     loading an extremely large table. See the
+#'     [xMart4 API documentation](https://portal-uat.who.int/xmart4/docs/xmart_api/use_API.html)
+#'     for details on the three options.
 #' @param auth_type Type of authorization to use for the token authorization. For
 #'     "wims", the default, this uses the WHO WIMS system to authenticate, with
 #'     the AzureAuth package on the backend. If "client", it uses an AzureAD client
 #'     setup.
 #' @param full_table Logical, whether or not to load all the rows in a specified table,
-#'     defaults to `TRUE`. The xMart4 API limits calls to 10,000 rows at a time, so
+#'     defaults to `TRUE`. If no format is provided, the xMart4 API limits calls to 100,000 rows at a time, so
 #'     if `full_table == TRUE`, the function automatically repeats the API call to
-#'     extract all rows within the xMart4 table.
+#'     extract all rows within the xMart4 table. Ignored if `format="streaming"` or
+#'     `format="csv"`.
 #' @param token Access token for xMart4 server. If NULL (the default), the package automatically creates and manages access for the user if Azure client ID and secret set up properly. See `vignette("token_setup")` for instructions and details.
 #'
 #' @return A data frame.
@@ -25,6 +32,7 @@ xmart4_api <- function(mart,
                        xmart_server = c("UAT", "PROD"),
                        top = NULL,
                        query = NULL,
+                       format = c("none", "csv", "streaming"),
                        full_table = TRUE,
                        auth_type = c("wims", "client"),
                        token = NULL) {
@@ -33,11 +41,13 @@ xmart4_api <- function(mart,
   xmart_server <- rlang::arg_match(xmart_server)
   assert_query(query)
   assert_top(top)
+  format <- rlang::arg_match(format)
   auth_type <- rlang::arg_match(auth_type)
 
   query <- modify_query(query)
   t_q <- join_top_query(top, query)
   t_q <- join_tq_skip(t_q, full_table)
+  t_q <- join_format(t_q, format)
   token <- check_raw_token(token, auth_type, xmart_server)
   url <- xmart_url(mart,
                    table,
@@ -46,7 +56,8 @@ xmart4_api <- function(mart,
   xmart4_get(url,
              t_q,
              token,
-             full_table)
+             full_table,
+             format)
 }
 
 #' @noRd
@@ -87,6 +98,7 @@ join_top_query <- function(top, query) {
   paste(x, collapse = "&")
 }
 
+#' @noRd
 join_tq_skip <- function(tq, full_table) {
   if (full_table) {
     x <- c("$skip=0", tq)
@@ -97,22 +109,51 @@ join_tq_skip <- function(tq, full_table) {
 }
 
 #' @noRd
-xmart4_get <- function(url, t_q, token, full_table) {
+join_format <- function(tq, format) {
+  if (format == "csv") {
+    tq <- paste(c(tq, "$format=csv"), collapse = "&")
+  } else if (format == "streaming") {
+    tq <- paste(c(tq, "$format=streaming"), collapse = "&")
+  }
+  tq
+}
+
+#' @noRd
+xmart4_get <- function(url, t_q, token, full_table, format) {
   resp <- httr::GET(httr::modify_url(url),
                     token_header(token),
                     ua,
                     query = t_q)
   assert_status_code(resp)
-  assert_json(resp)
-  parsed <- httr::content(resp,
-                          as = "parsed",
-                          type = "application/json")
-  assert_content(parsed)
-  df <- parsed_to_df(parsed)
-  next_link <- parsed[["@odata.nextLink"]]
+  assert_resp(resp, format)
+
+  rl <- xmart4_parse(resp, format)
+  df <- rl[[1]]
+  next_link <- rl[[2]]
+
   if (full_table & !is.null(next_link)) {
     params <- unlist(stringr::str_match_all(next_link, "(.+?)\\?(\\$.+)"))
     df <- dplyr::bind_rows(df, xmart4_get(params[2], params[3], token, full_table))
   }
   df
+}
+
+#' @noRd
+xmart4_parse <- function(resp, format) {
+  if (format %in% c("streaming", "none")) {
+    parsed <- httr::content(resp,
+                            as = "parsed",
+                            type = "application/json",
+                            encoding = "UTF-8")
+    assert_content(parsed)
+    df <- parsed_to_df(parsed)
+    next_link <- parsed[["@odata.nextLink"]]
+  } else {
+    df <- httr::content(resp,
+                        as = "parsed",
+                        type = "text/csv",
+                        encoding = "UTF-8")
+    next_link <- NULL
+  }
+  list(df, next_link)
 }